requirements.txt
azure-cosmos==4.2.0
Flask==1.1.2